// clients/sdk/src/generated/dac/accounts/networkConfig.ts
export type NetworkConfig = {
  discriminator: ReadonlyUint8Array;
  authority: Address;
  cidConfig: string;
  genesisHash: ReadonlyUint8Array;
  agentCount: bigint;
  goalCount: bigint;
  taskCount: bigint;
  validatorNodeCount: bigint;
  computeNodeCount: bigint;
  approvedCodeMeasurements: Array<CodeMeasurement>;
  bump: number;
};